requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
httpx>=0.28.0
python-multipart>=0.0.20
pdfservices-sdk>=4.2.0
python-dotenv>=1.0.1
